requirements.txt
asgiref==3.8.1
Django==4.2.7
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-gis==1.0
psycopg2-binary==2.9.7
pytz==2025.2
sqlparse==0.5.3
typing_extensions==4.14.0
tzdata==2025.2
